Bug 1197186 - Move project configuration to a separate folder
