+++ .gitignore
@@ -33,8 +33,8 @@ nosetests.xml
 
 #local settings
 deployment/update/commander_settings.py
-treeherder/settings/local.py
-treeherder/settings/pycharm_django_settings.py
+treeherder/config/settings/local.py
+treeherder/config/settings/pycharm_django_settings.py
 ui/js/config/local.conf.js
 
 htmlcov/
