+++ tests/test_utils.py
@@ -7,7 +7,7 @@ from tests.sampledata import SampleData
 from treeherder.client import TreeherderAuth, TreeherderClient
 from treeherder.etl.oauth_utils import OAuthCredentials
 from treeherder.model.derived.refdata import RefDataManager
-from treeherder.webapp.wsgi import application
+from treeherder.config.wsgi import application
 
 
 def post_collection(
