+++ tests/webapp/conftest.py
@@ -5,7 +5,7 @@ from webtest.app import TestApp
 
 from tests.sample_data_generator import job_data
 from treeherder.model.derived import ArtifactsModel
-from treeherder.webapp import wsgi
+from treeherder.config import wsgi
 
 
 @pytest.fixture
