+++ treeherder/config/wsgi.py
@@ -12,12 +12,12 @@ import os
 # if running multiple sites in the same mod_wsgi process. To fix this, use
 # mod_wsgi daemon mode with each site in its own daemon process, or use
 # os.environ["DJANGO_SETTINGS_MODULE"] = "webapp.settings"
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "treeherder.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "treeherder.config.settings")
 
 from django.core.cache.backends.memcached import BaseMemcachedCache
 from django.core.wsgi import get_wsgi_application
 
-from treeherder.webapp.whitenoise_custom import CustomWhiteNoise
+from treeherder.config.whitenoise_custom import CustomWhiteNoise
 
 # This application object is used by any WSGI server configured to use this
 # file. This includes Django's development server, if the WSGI_APPLICATION
